+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 * Path generation for image extraction in CLI command `hexapdf images`
 * Handling of certain invalid PDFs where the generation number for object
   identifiers don't match their cross-reference section value
+* AES 256bit encryption to include unnecessary field /Length in encryption
+  dictionary to work around buggy PDF libraries
 
 
 ## 1.5.0 - 2025-12-08